refactor server-store module

+++ service-store/contract/src/main/java/com/stayrascal/cloud/store/contract/client/LocationServiceClient.java
@@ -8,9 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
-@FeignClient(
-        name = "service-store"
-)
+@FeignClient(name = "service-store")
 public interface LocationServiceClient {
     @RequestMapping(
             method = {RequestMethod.GET},
@@ -18,7 +16,7 @@ public interface LocationServiceClient {
             consumes = {"application/json"}
     )
     PageResult list(@PathVariable("latitude") Double var1, @PathVariable("longitude") Double var2,
-                    @RequestParam("distance") Integer var3, @RequestParam("province_id") Long var4,
+                    @RequestParam("distance") Double var3, @RequestParam("province_id") Long var4,
                     @RequestParam("city_id") Long var5, @RequestParam("search_content") String var6,
                     @RequestParam("sort_type") String var7, @RequestParam("sort_by") String var8,
                     @RequestParam("page_size") Integer var9, @RequestParam("page_index") Integer var10);
